Use cobra for configuration (#120)
* first pass - support all options via cobra for command line usage

* cleanup, compatability with wrappers

* update changelog

* lint

* update orb

* remove delimiters from orb

* update version string

* length guard

* first pass - move search code to a search package

* Revert "first pass - move search code to a search package"

This reverts commit 49e823f4198905fc288ca1865c9caae0bd308289.

* rename declarations.go to flags.go

* support case insensitive filepattern declaration

+++ internal/ld/ld.go
@@ -395,7 +395,7 @@ type BranchCollection struct {
 type BranchRep struct {
 	Name             string              `json:"name"`
 	Head             string              `json:"head"`
-	UpdateSequenceId *int64              `json:"updateSequenceId,omitempty"`
+	UpdateSequenceId *int                `json:"updateSequenceId,omitempty"`
 	SyncTime         int64               `json:"syncTime"`
 	References       []ReferenceHunksRep `json:"references,omitempty"`
 }
